requirements.txt
twitchio==2.7.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
